Initial checkin of working feature
Status:
* Working [dev] web pack, npm, etc
* Several example/tests/debugging views (idiosyncratic)
* Wang definitions, initial grid implementation
* Edge tilesets
* Blob tileset

Corner tilesets are (logically) broken in this checkin, because they
work differently from the others. Edge tilesets work with *internal*
data and are effectively a projection of the wangid.

Corner tilesets aren't; they require identifying the data with its
upper-left corner, and using that to derive wangids (which will also
result in drawing content in tiles which are "off").

A followup diff will provide cleanup and an alternate corner
implementation.

The documentation also drifted, as implementation occurred.