[Major refactor] Make corner indices work.
Remove the blob indexer in favor of an explicit class to construct blob
textures.
Modify the Id class to support multiple distinct mechanisms for
calculating wangIds and tile mappings.

Remove the Grid class (and concept), split instead into a small number
of callbacks and the Pointwise class, which is explicitly about 2d
geometry and callbacks.